requirements.txt
paramiko
prometheus-client
cryptography
aiohttp
requests
protobuf
python-snappy
pyyaml
